+++ bindings/nodejs/types/account.ts
@@ -55,6 +55,7 @@ export interface NativeTokenBalance {
     tokenId: string;
     total: HexEncodedAmount;
     available: HexEncodedAmount;
+    metadata?: string;
 }
 
 /** Sync options for an account */
@@ -90,6 +91,8 @@ export interface AccountSyncOptions {
     /** Specifies if only basic outputs with an AddressUnlockCondition alone should be synced, will overwrite
      * `syncAliasesAndNfts`. Default: false. */
     syncOnlyMostBasicOutputs?: boolean;
+    /** Sync native token foundries, so their metadata can be returned in the balance. Default: false. */
+    syncNativeTokenFoundries?: boolean;
 }
 
 /** The account object */